refactor: change folder name from "const" to "consts"; transport mocks files in folder "mocks"
