+++ src/components/PrivateRoutes/PrivateRoutes.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from 'react-router-dom';
-import { AppRoute } from '../../const/pageRoutes';
+import { AppRoute } from '../../consts/pageRoutes';
 
 export function PrivateRoutes(): JSX.Element {
   const hasAccess = false;
