+++ src/components/PublicRoutes/PublicRoutes.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from 'react-router-dom';
-import { AppRoute } from '../../const/pageRoutes';
+import { AppRoute } from '../../consts/pageRoutes';
 
 export function PublicRoutes() {
   const hasAccess = false;
