+++ src/consts/pageRoutes.ts
@@ -1,0 +1,7 @@
+export const AppRoute = {
+  Main: '/',
+  Login: '/login',
+  Favorites: '/favorites',
+  Offer: '/offer',
+  OfferId: '/offer/:id',
+} as const;
