+++ src/index.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import '../markup/css/main.css';
 import App from './App';
-import { FAVORITE_PLACES } from './const/favoritePlaces';
-import { PLACE_CARDS } from './const/placeVariant';
+import { FAVORITE_PLACES } from './mocks/favoritePlaces';
+import { PLACE_CARDS } from './mocks/placeVariant';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
