+++ src/mocks/favoritePlaces.ts
